// src/types/OrdersHistory/RatingOrderHistoryCard.ts
export interface IRatingOrderHistoryCard {
  id: number;
  id_user: number;
  id_tovar: any;
  name: string;
  text: string;
  dat1: string;
  moder: number;
  ocenka: number;
  is_nal: any;
  naim_tovar: any;
  naim_city: string;
  url_tovar: string;
  all_otz: any;
  tel: any;
  type: number;
  carusel: any;
  service_type: string;
  id_zakaz: number;
  dostoinsva: any;
  nedostatki: any;
  anonim: any;
}